Repeated XOR cracking works
At least from what I tested, it works.

The algorithm weighs the decrypted chars differently.
A space is worth a lot, since those tend to be in encrypted
messages. Letters are next on the ladder, then numbers, and
then punctuation symbols.

The algorithm cycles through the printable characters to see
which character has the best score. This character will probably
be part of the key.